refactor: split config models into models.py (#86)
models.py takes the Pydantic models — User, OAuthClient, Settings with
its validators and the profile-derived protocol properties (#68) — plus
the YAML shape coercers. config.py keeps env-var expansion, the
ConfigManager and the singletons, and re-exports the models so every
existing import (routes, services, MCP, tests) keeps working unchanged.

Authentication stays on ConfigManager, per the issue's escape hatch:
authenticate/hash_password/check_client are 40 lines operating purely
on config state, and every caller treats them as config operations — a
separate service would be indirection for its own sake.
